// server/types/transferwise.ts
export type Quote = {
  id?: number;
  source: string;
  target: string;
  sourceAmount: number;
  targetAmount: number;
  type: string;
  rate: number;
  createdTime: string;
  createdByUserId: number;
  profile: number;
  rateType: string;
  deliveryEstimate: string;
  fee: number;
  allowedProfileTypes: Array<'PERSONAL' | 'BUSINESS'>;
  guaranteedTargetAmount: boolean;
  ofSourceAmount: boolean;
};

export type RecipientAccount = {
  id?: number;
  currency: string;
  type: string;
  accountHolderName: string;
  legalType: 'PRIVATE' | 'BUSINESS';
  details: {
    address?: string;
    email?: string;
    accountNumber?: string;
    sortCode?: string;
    abartn?: string;
    accountType?: string;
    bankgiroNumber?: string;
    ifscCode?: string;
    bsbCode?: string;
    institutionNumber?: string;
    transitNumber?: string;
    phoneNumber?: string;
    bankCode?: string;
    russiaRegion?: string;
    routingNumber?: string;
    branchCode?: string;
    cpf?: string;
    cardNumber?: string;
    idType?: string;
    idNumber?: string;
    idCountryIso3?: string;
    idValidFrom?: string;
    idValidTo?: string;
    clabe?: string;
    swiftCode?: string;
    dateOfBirth?: string;
    clearingNumber?: string;
    bankName?: string;
    branchName?: string;
    businessNumber?: string;
    province?: string;
    city?: string;
    rut?: string;
    token?: string;
    cnpj?: string;
    payinReference?: string;
    pspReference?: string;
    orderId?: string;
    idDocumentType?: string;
    idDocumentNumber?: string;
    targetProfile?: string;
    targetUserId?: string;
    taxId?: string;
    job?: string;
    nationality?: string;
    interacAccount?: string;
    bban?: string;
    iban?: string;
    bic?: string;
    IBAN?: string;
    BIC?: string;
  };
};